+++ src/modules/golang.rs
@@ -58,6 +58,105 @@ fn format_go_version(go_stdout: &str) -> Option<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::modules::utils::test::render_module;
+    use ansi_term::Color;
+    use std::fs::{self, File};
+    use std::io;
+    use tempfile;
+
+    #[test]
+    fn folder_without_go_files() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = None;
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_go_file() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("main.go"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_go_mod() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("go.mod"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_go_sum() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("go.sum"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_godeps() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        let godeps = dir.path().join("Godeps");
+        fs::create_dir_all(&godeps)?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_glide_yaml() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("glide.yaml"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_gopkg_yml() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("Gopkg.yml"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+    #[test]
+    fn folder_with_gopkg_lock() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("Gopkg.lock"))?.sync_all()?;
+
+        let actual = render_module("golang", dir.path());
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🐹 v1.12.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
 
     #[test]
     fn test_format_go_version() {
