+++ tests/testsuite/main.rs
@@ -11,7 +11,6 @@ mod git_branch;
 mod git_commit;
 mod git_state;
 mod git_status;
-mod golang;
 mod hg_branch;
 mod hostname;
 mod jobs;
